+++ packages/wonder-blocks-tooltip/src/components/tooltip-popper.tsx
@@ -54,8 +54,8 @@ export default class TooltipPopper extends React.Component<Props> {
             placement,
             style: {
                 // NOTE(jeresig): We can't just use `popperProps.style` here
-                // as the Flow type doesn't match Aphrodite's CSS flow props
-                // (as it doesn't camelCase props). So we just copy over the
+                // as the TypeScript type doesn't match Aphrodite's CSS TypeScript
+                // props (as it doesn't camelCase props). So we just copy over the
                 // props that we need, instead.
                 top: popperProps.style.top,
                 left: popperProps.style.left,
@@ -75,7 +75,6 @@ export default class TooltipPopper extends React.Component<Props> {
             updateTailRef: this._tailRefTracker.updateRef,
             isReferenceHidden: popperProps.isReferenceHidden,
         } as const;
-        // @ts-expect-error [FEI-5019] - TS2345 - Argument of type '{ readonly placement: Placement; readonly style: { readonly top: Property.Top<string | number> | undefined; readonly left: Property.Left<string | number> | undefined; readonly bottom: Property.Bottom<...> | undefined; readonly right: Property.Right<...> | undefined; readonly position: Property.Position | undefined; ...' is not assignable to parameter of type 'PopperElementProps'.
         return children(bubbleProps);
     }