+++ packages/wonder-blocks-tooltip/src/util/types.ts
@@ -1,15 +1,17 @@
 import * as React from "react";
 
+import type {CSSProperties} from "aphrodite";
+
 export type getRefFn = (
     arg1?: React.Component<any> | Element | null | undefined,
 ) => void;
 
 export type Offset = {
-    bottom: string | null | undefined;
-    top: string | null | undefined;
-    left: string | null | undefined;
-    right: string | null | undefined;
-    transform: string | null | undefined;
+    bottom: CSSProperties["bottom"];
+    top: CSSProperties["top"];
+    left: CSSProperties["left"];
+    right: CSSProperties["right"];
+    transform: CSSProperties["transform"];
 };
 
 export type Placement =