+++ packages/wonder-blocks-core/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @khanacademy/wonder-blocks-core
 
+## 4.8.0
+
+### Minor Changes
+
+-   873f4a14: Update aphrodite lib def and StyleType
+
+### Patch Changes
+
+-   d816af08: Update build and test configs use TypeScript
+-   3891f544: Update babel config to include plugins that Storybook needed
+-   0d28bb1c: Configured TypeScript
+-   3d05f764: Fix HOCs and other type errors
+-   c2ec4902: Update eslint configuration, fix lint
+-   2983c05b: Include 'types' field in package.json
+-   77ff6a66: Generate Flow types from TypeScript types
+-   ec8d4b7f: Fix miscellaneous TypeScript errors
+
 ## 4.7.0
 
 ### Minor Changes