+++ packages/wonder-blocks-testing/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @khanacademy/wonder-blocks-testing
 
+## 7.1.11
+
+### Patch Changes
+
+-   873f4a14: Update type narrowing logic to work better with TypeScript
+-   d816af08: Update build and test configs use TypeScript
+-   3891f544: Update babel config to include plugins that Storybook needed
+-   3813715d: Update wonder-stuff dependencies (non-functional changes)
+-   0d28bb1c: Configured TypeScript
+-   3d05f764: Fix HOCs and other type errors
+-   c2ec4902: Update eslint configuration, fix lint
+-   2983c05b: Include 'types' field in package.json
+-   77ff6a66: Generate Flow types from TypeScript types
+-   ec8d4b7f: Fix miscellaneous TypeScript errors
+-   Updated dependencies [d816af08]
+-   Updated dependencies [3891f544]
+-   Updated dependencies [3813715d]
+-   Updated dependencies [0d28bb1c]
+-   Updated dependencies [3d05f764]
+-   Updated dependencies [c2ec4902]
+-   Updated dependencies [2983c05b]
+-   Updated dependencies [77ff6a66]
+-   Updated dependencies [ec8d4b7f]
+    -   @khanacademy/wonder-blocks-data@10.1.1
+
 ## 7.1.10
 
 ### Patch Changes